Atualizacao 14022024

+++ app/model/address.py
@@ -28,7 +28,7 @@ class ModelAddress(db.Model):
         nullable=False
     )
     country_id = db.Column(
-        SMALLINT(unsigned=True),
+        INTEGER(unsigned=True),
         db.ForeignKey('countries.id', onupdate='CASCADE'),
         nullable=False
     )
